requirements.txt
fastapi[standard]==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
bcrypt==4.1.2
PyJWT==2.8.0
python-decouple==3.8
pydantic[email]==2.6.4
passlib[bcrypt]